# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Orange Money",
        "description": "API complète pour le système de paiement Orange Money avec authentification OTP et mot de passe",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/send-otp": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Envoyer un OTP par SMS pour première authentification",
                "description": "Envoie un code OTP de 6 chiffres par SMS au numéro fourni pour la première connexion. L'OTP est valide pendant 5 minutes.",
                "operationId": "40f7591016f0d0a5ce412e71fee1f351",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone sénégalais au format international",
                                        "type": "string",
                                        "format": "phone",
                                        "example": "+221771234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP envoyé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OTP envoyé avec succès"
                                        },
                                        "otp": {
                                            "description": "Code OTP généré",
                                            "type": "string",
                                            "example": "123456"
                                        },
                                        "sms_sent": {
                                            "description": "Indique si le SMS a été envoyé avec succès",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur d'envoi ou numéro invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors de l'envoi de l'OTP"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de requête invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/verify-otp": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier l'OTP et obtenir un token d'accès",
                "description": "Vérifie l'OTP fourni et retourne un token d'accès Bearer si l'authentification réussit. Marque le compte comme vérifié par OTP.",
                "operationId": "7809ddcd47cab2525ccd43d1c2031bfd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "otp"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone utilisé pour l'envoi de l'OTP",
                                        "type": "string",
                                        "format": "phone",
                                        "example": "+221771234567"
                                    },
                                    "otp": {
                                        "description": "Code OTP de 6 chiffres reçu par SMS",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "description": "Durée de validité du token en secondes",
                                            "type": "integer",
                                            "example": 31536000
                                        },
                                        "access_token": {
                                            "description": "Token d'accès Bearer",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "OTP invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OTP invalide ou expiré"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de requête invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion avec mot de passe après vérification OTP",
                "description": "Authentifie l'utilisateur avec son numéro de téléphone et mot de passe après vérification OTP préalable.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "password"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone utilisé pour la vérification OTP",
                                        "type": "string",
                                        "format": "phone",
                                        "example": "+221771234567"
                                    },
                                    "password": {
                                        "description": "Mot de passe du compte",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "expires_in": {
                                            "description": "Durée de validité du token en secondes",
                                            "type": "integer",
                                            "example": 31536000
                                        },
                                        "access_token": {
                                            "description": "Token d'accès Bearer",
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "nom": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "telephone": {
                                                    "type": "string"
                                                },
                                                "type_compte": {
                                                    "type": "string"
                                                },
                                                "statut_compte": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentification échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Numéro de téléphone ou mot de passe invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de requête invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token d'accès Bearer obtenu via OTP ou authentification par mot de passe",
                "scheme": "bearer"
            }
        }
    },
...
}